workers/multi-globals/entry.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Multiple globals for Worker constructor: parsing scriptURL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe src="incumbent/incumbent.html"></iframe>
<button onclick="frames[0].hello()" id="button">Hello</button>

<script>
async_test((t) => {
  onload = () => {
    document.getElementById('button').click();
  };
  onmessage = t.step_func_done(e => {
    assert_equals(e.data, 'current');
  });
});
</script>
